requirements.txt
Django==5.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.10.0
python-dotenv==1.0.0
paypalrestsdk==1.13.1
gunicorn==21.2.0
whitenoise==6.6.0
